Script for scraping Instagram posts
This commit introduces a Python script utilizing the Instaloader library to scrape the first 'num_posts' posts from a specified Instagram user's profile. The script creates an instance of Instaloader, fetches the profile object, and downloads the posts to a target directory. It also handles exceptions, printing error messages when encountered. Note: Make sure to install the Instaloader library before running the script.